sso_be/requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<7.0
PyJWT>=2.8,<3.0
python-ldap>=3.4,<4.0
exchangelib>=5.0,<6.0
pywin32>=300
